# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "北冥有声小说",
    "description": "这是SWAGGER_2生成的接口文档",
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "用户查询列表",
        "operationId": "findListUsingGET",
        "parameters": [
          {
            "name": "role_code",
            "in": "query",
            "description": "role_code",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "real_name",
            "in": "query",
            "description": "real_name",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageIndex",
            "in": "query",
            "description": "pageIndex",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserQueryRes"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "创建用户",
        "operationId": "createUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "正常返回",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                }
              }
            }
          },
          "201": {
            "description": "用户已创建"
          },
          "401": {
            "description": "用户未登陆"
          },
          "403": {
            "description": "没有权限访问"
          },
          "404": {
            "description": "资源不存在"
          },
          "502": {
            "description": "系统错误"
          }
        }
      }
    },
...
    "/users/{userID}/password": {
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "修改密码",
        "operationId": "changePasswordUsingPUT",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "用户Id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "password"
                ],
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "description": "密码"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "用户管理"
        ],
        "summary": "重置密码",
        "operationId": "resetPasswordUsingDELETE",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "用户Id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ResponseResult": {
        "title": "ResponseResult",
        "type": "object",
        "properties": {
          "errCode": {
            "type": "string",
            "description": "错误代码-成功时返回:000",
            "example": "000"
          },
          "httpStats": {
            "type": "integer",
            "description": "http状态码",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "错误描述-成功时返回:ok",
            "example": "ok"
          },
          "payload": {
            "type": "object",
            "description": "返回结果对象"
          },
          "timestamp": {
            "type": "integer",
            "description": "执行时间",
            "format": "int64",
            "example": 1594954108020
          }
        },
        "description": "下行结果对象"
      },
...
      "UserCreateReq": {
        "title": "UserCreateReq",
        "type": "object",
        "properties": {
          "real_name": {
            "type": "string",
            "description": "姓名",
            "example": "陈凡"
          },
          "role_code": {
            "type": "array",
            "description": "角色代码",
            "example": "admin",
            "items": {
              "type": "string"
            }
          },
          "user_name": {
            "type": "string",
            "description": "账号",
            "example": "chengfan"
          }
        },
        "description": "用户创建上行"
      },
...
      "UserQueryRes": {
        "title": "UserQueryRes",
        "type": "object",
        "properties": {
          "lock": {
            "type": "boolean",
            "description": "是否锁定",
            "example": false
          },
          "real_name": {
            "type": "string",
            "description": "姓名",
            "example": "陈凡"
          },
          "role_code": {
            "type": "string",
            "description": "角色代码",
            "example": "admin"
          },
          "user_id": {
            "type": "string",
            "description": "用户id",
            "example": "adc8d0b4-5bee-41b9-a88a-599b6323ba8f"
          },
          "user_name": {
            "type": "string",
            "description": "账号",
            "example": "chengfan"
          }
        },
        "description": "用户查询下行"
      }
    }
  }
}